Changes during the tournament
During the tournament we realized that we forgot it include a selector option for the auto that would only get the skystone so we added an option for that. We also programmed a foundation auto specifically in accordance to our alliance partner, including a delay and extra movement to let them pass through the bridge.